Lightweight comboboxes
After this change, the combo boxes containing long lists
of club members, on the daily sheet, should scale:
the page should stay responsive even if many tenths of
launches are created in a session.

This is achieved having only one long selector for each
list, whilst previously each combo box was listing all
the possible options, and it was also doing it twice,
making the DOM grow to a large large number of nodes very
quickly.

- delete unused createDropDownList function.
- refresh adapted to the new situation. Note change to DailySheet
- two centralised "templates" (actually, singleton instances)
- reimplemented customised selectpicker (in the addTo functions)
- important note: ChargesSelect and CrewSelect are very similar but
  slightly different. It could be an idea to refactor the common code.